FEとの連携追加とCreativeEnergyCubeの追加(電力を吸収するけど無限に取り出せる)

+++ src/main/java/com/github/mrbestcreator/magicalmechanics/api/block/EnergyDef.java
@@ -1,8 +1,8 @@
 package com.github.mrbestcreator.magicalmechanics.api.block;
 
 public class EnergyDef {
-    public int capacity;
-    public int maxInput;
-    public int maxOutput;
+    public long capacity;
+    public long maxInput;
+    public long maxOutput;
     public int usagePerTick;
 }
